requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
supabase==2.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5 